Cleaning up alpaca-formatter-in-alpaca
It's still not doing anything helpful but I've started breaking things
down a bit and a number of the fixes over the last couple of days have
smoothed things out enough to show where I'm going.  Contents of the
folder `src/lib` are helper functions, not necessarily a good start on
anything approaching a standard library.